+++ govtool/frontend/src/i18n/index.ts
@@ -1,11 +1,13 @@
 import i18n from "i18next";
 import { initReactI18next } from "react-i18next";
 
-import { en } from "./locales/en";
+import en from "./locales/en.json";
 
 i18n.use(initReactI18next).init({
   resources: {
-    en,
+    en: {
+      translation: en,
+    },
   },
   fallbackLng: "en",
   interpolation: {