+++ govtool/frontend/src/services/requests/getProposal.ts
@@ -1,11 +1,18 @@
 import { VotedProposal, VotedProposalDTO } from "@/models";
+import { decodeCIP129Identifier, mapDtoToProposal } from "@/utils";
+
 import { API } from "../API";
-import { mapDtoToProposal } from "@/utils";
 
 export const getProposal = async (
   proposalId: string,
   drepId?: string,
 ): Promise<VotedProposal> => {
+  const isCIP129Identifier = proposalId.includes("gov_action");
+  if (isCIP129Identifier) {
+    const { txID } = decodeCIP129Identifier(proposalId);
+    proposalId = txID;
+  }
+
   const encodedHash = encodeURIComponent(proposalId);
 
   const { data } = await API.get<VotedProposalDTO>(